backend/requirements.txt
# ─────────────────────────────────────────
# Core Framework
# ─────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# ─────────────────────────────────────────
# LangGraph / LangChain
# ─────────────────────────────────────────
langgraph==0.2.60
langchain==0.3.12
langchain-openai==0.2.12
langchain-core==0.3.27

# ─────────────────────────────────────────
# OpenAI
# ─────────────────────────────────────────
openai==1.58.1

# ─────────────────────────────────────────
# Database
# ─────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ─────────────────────────────────────────
# Redis
# ─────────────────────────────────────────
redis==5.2.1
hiredis==3.0.0

# ─────────────────────────────────────────
# Security
# ─────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# ─────────────────────────────────────────
# Utilities
# ─────────────────────────────────────────
structlog==24.4.0
python-dotenv==1.0.1
httpx==0.28.1

# ─────────────────────────────────────────
# Testing
# ─────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
